Перенес методы из DiagnosticHelper в Trees
 - getPreviousNode
 - getRootParent
 - findErrorNode
