+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/utils/DiagnosticHelper.java
@@ -78,51 +78,4 @@ public final class DiagnosticHelper {
     return "ФиксированнаяСтруктура".equalsIgnoreCase(tnc.getText()) || "FixedStructure".equalsIgnoreCase(tnc.getText());
   }
 
-  /**
-   * Проверяет среди дочерних элементов наличие ноды с ошибкой
-   * @return true - если есть нода с ошибкой
-   */
-  public static boolean findErrorNode(ParseTree tnc) {
-
-    if (tnc instanceof BSLParserRuleContext) {
-      if (((BSLParserRuleContext) tnc).exception != null) {
-        return true;
-      }
-
-      for (int i = 0; i < tnc.getChildCount(); i++) {
-        if (findErrorNode(tnc.getChild(i))) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Выполняет поиск предыдущей ноды нужного типа
-   * @param parent - родительская нода, среди дочерних которой производится поиск
-   * @param tnc - нода, для которой ищем предыдущую
-   * @param ruleindex - BSLParser.RULE_*
-   * @return tnc - если предыдущая нода не найдена, вернет текущую
-   */
-  public static ParseTree getPreviousNode(ParseTree parent, ParseTree tnc, int ruleindex) {
-    List<ParseTree> statements = new ArrayList<>(Trees.findAllRuleNodes(parent, ruleindex));
-
-    int pos = statements.indexOf(tnc);
-    if ((pos - 1) > -1) {
-      return statements.get(pos - 1);
-    }
-    return tnc;
-  }
-
-  /**
-   * Рекурсивно находит самого верхнего родителя текущей ноды
-   */
-  public static ParseTree getRootParent(ParseTree tnc) {
-    if(tnc.getParent() != null) {
-      return getRootParent(tnc.getParent());
-    }
-
-    return tnc;
-  }
 }
